Address PR feedback.
