+++ src/System.Text.JsonLab.Dynamic/System/Text/Json/JsonDynamicObject.cs
@@ -5,12 +5,11 @@ using System.Buffers;
 using System.Buffers.Text;
 using System.Collections.Generic;
 using System.Dynamic;
-using System.Reflection;
 using System.Text.Utf8;
 
 namespace System.Text.JsonLab
 {
-    public partial class JsonDynamicObject : DynamicObject, IBufferFormattable
+    public class JsonDynamicObject : DynamicObject, IBufferFormattable
     {
         private readonly Dictionary<JsonProperty, JsonValue> _properties;
 
@@ -19,78 +18,6 @@ namespace System.Text.JsonLab
         private JsonDynamicObject(Dictionary<JsonProperty, JsonValue> properties)
         {
             _properties = properties;
-        }
-
-        private static Dictionary<Type, Dictionary<int, PropertyInfoLinkedList>> TypeCache = new Dictionary<Type, Dictionary<int, PropertyInfoLinkedList>>();
-
-        public static T Deserialize<T>(ReadOnlySpan<byte> utf8)
-        {
-            if (!TypeCache.TryGetValue(typeof(T), out Dictionary<int, PropertyInfoLinkedList> dictionary))
-            {
-                dictionary = GetTypeMap(typeof(T));
-                TypeCache.Add(typeof(T), dictionary);
-            }
-
-            T instance = Deserializer<T>.Deserialize(utf8);
-
-            var reader = new JsonReader(utf8, SymbolTable.InvariantUtf8);
-            while (reader.Read())
-            {
-                switch (reader.TokenType)
-                {
-                    case JsonTokenType.PropertyName:
-                        int key = GetHashCode(reader.Value);
-                        reader.Read(); // Move to the value token
-                        JsonValueType type = reader.ValueType;
-                        switch (type)
-                        {
-                            case JsonValueType.String:
-                                PropertyInfo pi = GetPropertyInfo(dictionary, key, reader.Value);
-                                pi.SetValue(instance, new Utf8String(reader.Value));    // TODO: Use Ref.Emit instead of Reflection
-                                break;
-                            case JsonValueType.Object: // TODO: could this be lazy? Could this reuse the root JsonObject (which would store non-allocating JsonDom)?
-                                throw new NotImplementedException("object support not implemented yet.");
-                            case JsonValueType.True:
-                                pi = GetPropertyInfo(dictionary, key, reader.Value);
-                                pi.SetValue(instance, true);
-                                break;
-                            case JsonValueType.False:
-                                pi = GetPropertyInfo(dictionary, key, reader.Value);
-                                pi.SetValue(instance, false);
-                                break;
-                            case JsonValueType.Null:
-                                pi = GetPropertyInfo(dictionary, key, reader.Value);
-                                pi.SetValue(instance, null);
-                                break;
-                            case JsonValueType.Number:
-                                pi = GetPropertyInfo(dictionary, key, reader.Value);
-                                if (!Utf8Parser.TryParse(reader.Value, out int result, out _))
-                                {
-                                    throw new InvalidCastException();
-                                }
-                                pi.SetValue(instance, result);
-                                break;
-                            case JsonValueType.Array:
-                                throw new NotImplementedException("array support not implemented yet.");
-                            default:
-                                throw new NotSupportedException();
-                        }
-                        break;
-                    case JsonTokenType.StartObject:
-                        break;
-                    case JsonTokenType.EndObject:
-                        break;
-                    case JsonTokenType.StartArray:
-                        throw new NotImplementedException("array support not implemented yet.");
-                    case JsonTokenType.EndArray:
-                    case JsonTokenType.Value:
-                        break;
-                    default:
-                        throw new NotSupportedException();
-                }
-            }
-
-            return instance;
         }
 
         public static JsonDynamicObject Parse(ReadOnlySpan<byte> utf8, int expectedNumberOfProperties = -1)
@@ -289,67 +216,6 @@ namespace System.Text.JsonLab
         private static unsafe bool TryEncodeControlChar(SymbolTable symbolTable, byte value, Span<byte> buffer, out int written)
         {
             return symbolTable.TryEncode(value, buffer, out written);
-        }
-
-        private static int GetHashCode(ReadOnlySpan<byte> span)
-        {
-            int hash = 17;
-            foreach (byte element in span)
-            {
-                hash = hash * 31 + element;
-            }
-            return hash;
-        }
-
-        private static Dictionary<int, PropertyInfoLinkedList> GetTypeMap(Type type)
-        {
-            var dictionary = new Dictionary<int, PropertyInfoLinkedList>();
-            IEnumerable<PropertyInfo> properties = type.GetRuntimeProperties();
-            foreach (PropertyInfo pi in properties)
-            {
-                byte[] encoded = Encoding.UTF8.GetBytes(pi.Name);
-                int hashcode = GetHashCode(encoded);
-
-                if (dictionary.ContainsKey(hashcode))
-                {
-                    dictionary.TryGetValue(hashcode, out PropertyInfoLinkedList list);
-                    (byte[], PropertyInfo) data = (encoded, pi);
-                    list.Add(data);
-                }
-                else
-                {
-                    var list = new PropertyInfoLinkedList();
-                    (byte[], PropertyInfo) data = (encoded, pi);
-                    list.Add(data);
-                    dictionary.Add(hashcode, list);
-                }
-            }
-            return dictionary;
-        }
-
-        public static PropertyInfo GetPropertyInfo(Dictionary<int, PropertyInfoLinkedList> dictionary, int key, ReadOnlySpan<byte> span)
-        {
-            if (!dictionary.TryGetValue(key, out PropertyInfoLinkedList value))
-            {
-                throw new KeyNotFoundException();
-            }
-            PropertyInfoNode node = value.Head;
-            PropertyInfo pi = node.Value.propertyInfo;
-
-            // This should be a very rare occurrence (only if there is hash collision)
-            if (value.Count > 1)
-            {
-                while (node != null)
-                {
-                    if (span.SequenceEqual(node.Value.encodedName))
-                    {
-                        pi = node.Value.propertyInfo;
-                        break;
-                    }
-                    node = node.Next;
-                }
-            }
-            return pi;
         }
 
         struct JsonValue : IBufferFormattable
